+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 8. Added pacseek
 9. Added reflector config
 10. Made wal colors load first in shell
+11. Added statusbar, built with quickshell